More MSVC warnings fixes

+++ src/lib/hash/par_hash/par_hash.cpp
@@ -18,7 +18,7 @@ void Parallel::add_data(const uint8_t input[], size_t length)
 
 void Parallel::final_result(uint8_t out[])
    {
-   uint32_t offset = 0;
+   size_t offset = 0;
 
    for(auto&& hash : m_hashes)
       {
